# todolistfrontend/src/app/services/models/task.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { Users } from '../models/users';
export interface Task {
  completed?: boolean;
  dueDate?: string;
  id?: number;
  title?: string;
  user?: Users;
}