# todolistfrontend/src/app/services/models/users.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { GrantedAuthority } from '../models/granted-authority';
import { Task } from '../models/task';
export interface Users {
  accountNonExpired?: boolean;
  accountNonLocked?: boolean;
  authorities?: Array<GrantedAuthority>;
  credentialsNonExpired?: boolean;
  email?: string;
  enabled?: boolean;
  id?: number;
  role?: 'USER' | 'ADMIN';
  tasks?: Array<Task>;
  username?: string;
}