# todolistfrontend/src/openapi/openapi.json
{"openapi":"3.0.1","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"paths":{"/api/tasks/{id}":{"put":{"tags":["task-controller"],"operationId":"updateTask","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Task"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Task"}}}}}},"delete":{"tags":["task-controller"],"operationId":"deleteTask","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK"}}}},"/api/tasks":{"get":{"tags":["task-controller"],"operationId":"getAllTasksByUser","parameters":[{"name":"userId","in":"query","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Task"}}}}}}},"post":{"tags":["task-controller"],"operationId":"createTask","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Task"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Task"}}}}}}},"/api/auth/register":{"post":{"tags":["Authentication"],"operationId":"register","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AuthenticationResponse"}}}}}}},"/api/auth/refresh":{"post":{"tags":["Authentication"],"operationId":"refresh","responses":{"200":{"description":"OK"}}}},"/api/auth/auth":{"post":{"tags":["Authentication"],"operationId":"authentication","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthenticationRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AuthenticationResponse"}}}}}}},"/users/getuser":{"get":{"tags":["users-controller"],"operationId":"getuser","parameters":[{"name":"Username","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Users"}}}}}}},"/api/tasks/upcoming":{"get":{"tags":["task-controller"],"operationId":"getUpcomingTasks","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Task"}}}}}}}},"/api/tasks/today":{"get":{"tags":["task-controller"],"operationId":"getTodayTasks","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Task"}}}}}}}}},"components":{"schemas":{"GrantedAuthority":{"type":"object","properties":{"authority":{"type":"string"}}},"Task":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"title":{"type":"string"},"dueDate":{"type":"string","format":"date-time"},"completed":{"type":"boolean"},"user":{"$ref":"#/components/schemas/Users"}}},"Users":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"tasks":{"type":"array","items":{"$ref":"#/components/schemas/Task"}},"username":{"type":"string"},"email":{"type":"string"},"role":{"type":"string","enum":["USER","ADMIN"]},"enabled":{"type":"boolean"},"authorities":{"type":"array","items":{"$ref":"#/components/schemas/GrantedAuthority"}},"accountNonExpired":{"type":"boolean"},"accountNonLocked":{"type":"boolean"},"credentialsNonExpired":{"type":"boolean"}}},"RegisterRequest":{"type":"object","properties":{"username":{"type":"string"},"password":{"type":"string"},"email":{"type":"string"},"role":{"type":"string","enum":["USER","ADMIN"]}}},"AuthenticationResponse":{"type":"object","properties":{"accessToken":{"type":"string"},"refreshToken":{"type":"string"}}},"AuthenticationRequest":{"type":"object","properties":{"username":{"type":"string"},"password":{"type":"string"}}}}}}